Node adding cors and web Api

+++ BlockChain.Core/BlockChain.Core/ITransactionSigner.cs
@@ -4,7 +4,7 @@ namespace BlockChain.Core
 {
     public interface ITransactionSigner
     {
-        Transaction Sign(string privateKey, string recipientAddress, decimal amount, DateTime signDate);
+        Transaction Sign(string privateKey, string recipientAddress, ulong amount, DateTime signDate);
         string CalculateAddress(string privateKey);
     }
